requirements.txt
Pillow
torch>2
timm>0.9